Fix ShapeAndTypeBased and ShapeAndTypeAndDeviceBased cases
- Fix _inputs to use consistent float16 dtypes (was float32/float16 mix)
- Add _handle_zeros and _handle_ones to GraphTracer for TracingInt sizes
- Add _zeros_replacement_ctx and _ones_replacement_ctx patches
- Include new patches in _trace_replacement_ctx
- Add torch.device to is_not_tensor for safety
- Add test methods for both models using yobx-tracing with dynamic=True

Agent-Logs-Url: https://github.com/xadupre/yet-another-onnx-builder/sessions/29bfd980-f637-476c-89e7-cd7002428d53

+++ yobx/torch/testing/_model_eval_cases.py
@@ -1142,7 +1142,7 @@ class ShapeAndTypeBased(torch.nn.Module):
         return torch.zeros(new_shape, dtype=dtype)
 
     _inputs = [
-        (torch.rand((3, 4), dtype=torch.float32),),
+        (torch.rand((3, 4), dtype=torch.float16),),
         (torch.rand((5, 4), dtype=torch.float16),),
     ]
     _dynamic = {"x": {0: DIM("batch"), 1: DIM("seq")}}
@@ -1159,7 +1159,7 @@ class ShapeAndTypeAndDeviceBased(torch.nn.Module):
         return torch.zeros(new_shape, dtype=dtype, device=device)
 
     _inputs = [
-        (torch.rand((3, 4), dtype=torch.float32),),
+        (torch.rand((3, 4), dtype=torch.float16),),
         (torch.rand((5, 4), dtype=torch.float16),),
     ]
     _dynamic = {"x": {0: DIM("batch"), 1: DIM("seq")}}
